Tests for Classes of Generalized Ellipsoids

These functions provide tests for classes of generalized ellipsoids in the (U, D) representation, based on the numbers of positive, zero and infinite singular values in D.

They are included here mainly as computational definitions of the terms ‘bounded’ for an ellipsoid with finite extent, ‘fat’, for a bounded ellipsoid with non-empty interior, ‘flat’, for degenerate (singular) ellipsoids in \(R^p\) with empty interior.

Usage

isBounded(x, ...)

# S3 method for class 'gell'
isBounded(x, ...)

isFat(x, ...)

# S3 method for class 'gell'
isFat(x, ...)

isFlat(x, ...)

# S3 method for class 'gell'
isFlat(x, ...)

isUnbounded(x, ...)

# S3 method for class 'gell'
isUnbounded(x, ...)

Arguments

x

A class "gell" object

...

Other arguments, not used.

Value

TRUE or FALSE

References

Friendly, M., Monette, G. and Fox, J. (2013). Elliptical Insights: Understanding Statistical Methods through Elliptical Geometry. Statistical Science, 28(1), 1-39.

See also

Author

Georges Monette

Examples

# None yet